(Probably?) Eliminated mpd disconnect problems
MPD runs on a fairly short client timeout[1], so you actually
do *need* to handle disconnects reliably as a matter of course.

Since we have the Go library 'watcher' running, and it doesn't seem to
be affected by timeouts (probably handles it internally) we can have
the watcher connection open until an interesting event happens, at
which we actually connect to MPD and fetch the data associated with
the event.

As such, I've actually been able to remove parallelism, since it's not
needed.  A much simpler and more robust design.

[1] https://github.com/Mic92/python-mpd2/issues/31#issuecomment-31415591