Moving to go modules
I like Go but there are some parts of it that I'm not enamored with.
Having to make everything live under the GOPATH was one.  But it isn't
a problem as of 1.11 -- `go mod` allows you to vendor your
dependencies per-project and not be tied to a specific directory on
your disk just to make things work...

I followed [1] to get this set up.  I'm assuming go.mod and go.sum
should be added to source control, like with package.json or
composer.json/lock, etc etc.

I guess this is confirmation that this junk works with 1.11 as well :P
Tested (briefly) with 1.11.4 on Ubuntu 16.04.

[1] https://www.kablamo.com.au/blog-1/2018/12/10/just-tell-me-how-to-use-go-modules